+++ src/app/page.tsx
@@ -4,17 +4,7 @@ import Head from "next/head";
 import Image from "next/image";
 import { Inter } from "next/font/google";
 import styles from "../styles/Home.module.css";
-import {
-  Alert,
-  Box,
-  Button,
-  Container,
-  Link,
-  Snackbar,
-  Step,
-  StepLabel,
-  Stepper,
-} from "@mui/material";
+import { Alert, Box, Button, Container, Link, Snackbar } from "@mui/material";
 import NextLink from "next/link";
 const inter = Inter({ subsets: ["latin"] });
 
@@ -28,80 +18,23 @@ import { useRouter } from "next/router";
 import { join } from "path";
 import { useState } from "react";
 import StorageTable from "@/components/StorageTable";
-import { StorageTableRow } from "@/components/StorageTable";
-import EditTargetDialog from "@/components/EditTargetDialog";
-import HorizontalLinearStepper from "@/components/Stepper";
-
-import { invoke } from '@tauri-apps/api/core';
-
-interface StorageCredentials {
-  user_storage_name: string;
-  storage_type: string;
-  storage_name: string;
-  storage_account_id: string;
-  storage_access_key: string;
-  storage_secret_key: string;
-  timestamp: number;
-}
 
 export default function Home() {
   const [loading, setLoading] = useState(false);
-  const [activeStep, setActiveStep] = React.useState(0);
-  const [openDialog, setOpenDialog] = useState(false);
-  const [rows, setRows] = useState<StorageTableRow[]>([
-    { id: 0, name: "Dummy1", date: "Date added", connectionStatus: "Connection status" },
-    { id: 1, name: "Dummy2", date: "Date added", connectionStatus: "Connection status" },
-  ]);
 
-  const handleClickOpen = () => {
-    setOpenDialog(true);
-  };
-  const handleClose = () => {
-    setOpenDialog(false);
-  };
-  
+  const rows: any[] = [{ id: "1" }];
   const handleStart = () => {
     console.log("Starting something");
-    invoke<StorageCredentials[]>('get_all_storage_credentials').then((res) => {
-      setRows(res.map<StorageTableRow>((value, index) => {
-        return {
-          id: index,
-          name: value.user_storage_name,
-          date: new Date(value.timestamp * 1000).toLocaleString(),
-          connectionStatus: "Connected",
-        };
-      }));
-    });
   };
-  
   return (
     <main>
       <Container maxWidth="xl">
-        <Box sx={{ ml: 8, mr: 8 }}>
-          <Typography variant="h5" sx={{ mt: 5 }}>
-            {" "}
-            First page title {" "}
-          </Typography>
-          <Box sx={{ mt: 1 }}>
-            <HorizontalLinearStepper activeStep={activeStep} />
-          </Box>
-          <Box sx={{ mt: 4 }}>
-            <StorageTable rows={rows} />
-          </Box>
-          <Button
-            variant="contained"
-            sx={{ mt: 4 }}
-            disabled={loading}
-            onClick={handleStart}
-          >
-            Start something
-          </Button>
-          <Box></Box>
-          <Button variant="contained" sx={{ mt: 4 }} onClick={handleClickOpen}>
-            Open Dialog
-          </Button>
-        </Box>
-        <EditTargetDialog handleClose={handleClose} open={openDialog} />
+        <Typography
+            variant="h2"
+            sx={{ mt: 4, mb: 4, textAlign: "center" }}  
+            >
+            Welcome to the about page
+        </Typography>
       </Container>
     </main>
   );
